Submissions
Here are my submissions for the Burst Gun, Laser, Zolygon, and Dead Player Movement. For some reason GitHub only lets me submit to this branch. I will look into this on my end. Each submission and their scripts/prefabs are in their respective folders, as well as a test scene for each one.